requirements.txt
docx==0.2.4
langchain==1.0.3
langchain_community==0.4.1
langchain_core==1.0.3
langchain_huggingface==1.0.1
langgraph==1.0.2
pydantic==2.12.4
python-dotenv==1.2.1
python_docx==1.2.0
streamlit==1.48.1
